+++ base/types/gfsptask/upload.go
@@ -341,7 +341,7 @@ func (m *GfSpReplicatePieceTask) InitReplicatePieceTask(object *storagetypes.Obj
 	m.SetPriority(priority)
 	m.SetTimeout(timeout)
 	m.SetMaxRetry(retry)
-	m.SetNotAvailableSpIdx(-1)
+	m.NotAvailableSpIdx = -1
 }
 
 func (m *GfSpReplicatePieceTask) Key() coretask.TKey {
@@ -434,9 +434,7 @@ func (m *GfSpReplicatePieceTask) SetPriority(priority coretask.TPriority) {
 }
 
 func (m *GfSpReplicatePieceTask) SetNotAvailableSpIdx(idx int32) {
-	if atomic.LoadInt32(&m.NotAvailableSpIdx) != idx {
-		atomic.StoreInt32(&m.NotAvailableSpIdx, idx)
-	}
+	atomic.CompareAndSwapInt32(&m.NotAvailableSpIdx, -1, idx)
 }
 
 func (m *GfSpReplicatePieceTask) EstimateLimit() corercmgr.Limit {